+++ apps/cms/CHANGELOG.md
@@ -1,5 +1,19 @@
 # cms
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+  - @nospipi/database@1.0.0
+
 ## 0.1.1
 
 ### Patch Changes